+++ simpletransformers/classification/classification_model.py
@@ -72,6 +72,7 @@ from simpletransformers.config.global_args import global_args
 
 try:
     import wandb
+
     wandb_available = True
 except ImportError:
     wandb_available = False
@@ -110,12 +111,12 @@ class ClassificationModel:
             "flaubert": (FlaubertConfig, FlaubertForSequenceClassification, FlaubertTokenizer),
         }
 
-        if args and 'manual_seed' in args:
-            random.seed(args['manual_seed'])
-            np.random.seed(args['manual_seed'])
-            torch.manual_seed(args['manual_seed'])
-            if 'n_gpu' in args and args['n_gpu'] > 0:
-                torch.cuda.manual_seed_all(args['manual_seed'])
+        if args and "manual_seed" in args:
+            random.seed(args["manual_seed"])
+            np.random.seed(args["manual_seed"])
+            torch.manual_seed(args["manual_seed"])
+            if "n_gpu" in args and args["n_gpu"] > 0:
+                torch.cuda.manual_seed_all(args["manual_seed"])
 
         config_class, model_class, tokenizer_class = MODEL_CLASSES[model_type]
         if num_labels:
@@ -464,7 +465,9 @@ class ClassificationModel:
                         if best_eval_metric and args["early_stopping_metric_minimize"]:
                             if results[args["early_stopping_metric"]] - best_eval_metric < args["early_stopping_delta"]:
                                 best_eval_metric = results[args["early_stopping_metric"]]
-                                self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
+                                self._save_model(
+                                    args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                                )
                                 early_stopping_counter = 0
                             else:
                                 if args["use_early_stopping"]:
@@ -483,7 +486,9 @@ class ClassificationModel:
                         else:
                             if results[args["early_stopping_metric"]] - best_eval_metric > args["early_stopping_delta"]:
                                 best_eval_metric = results[args["early_stopping_metric"]]
-                                self._save_model(args["best_model_dir"], optimizer, scheduler, model=model, results=results)
+                                self._save_model(
+                                    args["best_model_dir"], optimizer, scheduler, model=model, results=results
+                                )
                                 early_stopping_counter = 0
                             else:
                                 if args["use_early_stopping"]:
@@ -728,8 +733,8 @@ class ClassificationModel:
         )
 
         if os.path.exists(cached_features_file) and (
-            (not args["reprocess_input_data"] and not no_cache) or (
-                mode == "dev" and args["use_cached_eval_features"] and not no_cache)
+            (not args["reprocess_input_data"] and not no_cache)
+            or (mode == "dev" and args["use_cached_eval_features"] and not no_cache)
         ):
             features = torch.load(cached_features_file)
             if verbose: