+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.23.0] - 2020-03-30
+
+### Added
+
+- Added support for training custom tokenizers.
+- Added improved support for training language models from scratch.
+- Added improved support for resuming training in classification, NER, and QnA tasks.
+
 ## [0.22.1] - 2020-03-19
 
 ### Added
@@ -478,9 +486,9 @@ Model checkpoint is now saved for all epochs again.
 
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 
-[0.22.1]: https://github.com/ThilinaRajapakse/simpletransformers/compare/9f093b7...HEAD
+[0.22.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/b5cf82c...HEAD
 
-[0.22.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/51fc7a3...9f093b7
+[0.22.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/51fc7a3...b5cf82c
 
 [0.21.5]: https://github.com/ThilinaRajapakse/simpletransformers/compare/27ff44e...51fc7a3